TrackPoint: Different sensitivity when not in mouse mode
Some values like sensitivity and PtS-threshold heavily depend on the
type of extension used. If it is more flexible, a much higher
sensitivity can be used. Also, limited clearance between caps can force
one to decrease it drastically for the first trackpoint detection or it
will activate unwantedly when hitting nearby caps from an angle or
slightly brushing the trackpoint extension when hitting keys behind it.

Sensitivity and threshold are configurable at runtime for tests.

[PS/2] low sensitivity for all boards.